Separate flood fill from the simulation client
